Started SerialPort (WIP)
